Refactor: extract Attack action into submodule
